From: Christoph Schroeder [mailto:christoph.schroeder@helmholtz-berlin.de] Sent: 27 July 2015 17:08 To: Mercado, Ronaldo (DLSLtd,RAL,TEC) Subject: Re: DLS EtherCAT Problem with reading SDOs
Hi Ronaldo,

I totally missed out, that there already is a template for SDOs. The main problem was the lack of the "_trig" record, which isn't mentioned in the documentation (http://controls.diamond.ac.uk/downloads/support/ethercat/4-3/documentation/doxygen/sdo_support.html).

I modified the Scanner to allow the reading of 32bit SDOs and couldn't observe the loss of the first 8bits because my values aren't that big, but there is a possible conversion problem since "sdocast_int32" returns a 32bit signed value while the most SDOs are unsigned.

I attached the changes I made to this mail. Thank you for the help.

Best regards,
Christoph